Move `ResultExt` to `result` module
