+++ packages/engine/lib/error/src/lib.rs
@@ -169,8 +169,8 @@ use core::{fmt, marker::PhantomData, mem::ManuallyDrop, panic::Location};
 
 use provider::Provider;
 
-pub use self::macros::*;
 use self::{frame::FrameRepr, report::ReportImpl};
+pub use self::{macros::*, result::ResultExt};
 
 /// Contains a [`Frame`] stack consisting of an original error, context information, and optionally
 /// a [`Backtrace`] and a [`SpanTrace`].
@@ -375,76 +375,6 @@ impl<C: Provider + fmt::Display + fmt::Debug + Send + Sync + 'static> Context fo
 //   Tracking issue: https://github.com/rust-lang/rust/issues/41517
 pub trait Message: fmt::Display + fmt::Debug + Send + Sync + 'static {}
 impl<M: fmt::Display + fmt::Debug + Send + Sync + 'static> Message for M {}
-
-/// Extension trait for [`Result`][core::result::Result] to provide context information on
-/// [`Report`]s.
-pub trait ResultExt<T> {
-    /// Type of the resulting context `C` inside of [`Report<C>`] when not providing a context.
-    type Context;
-
-    /// Adds new contextual message to the [`Frame`] stack of a [`Report`].
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use error::Result;
-    /// # fn load_resource(_: &User, _: &Resource) -> Result<()> { Ok(()) }
-    /// # struct User;
-    /// # struct Resource;
-    /// use error::ResultExt;
-    ///
-    /// # let user = User;
-    /// # let resource = Resource;
-    /// # #[allow(unused_variables)]
-    /// let resource = load_resource(&user, &resource).wrap_err("Could not load resource")?;
-    /// # Result::Ok(())
-    /// ```
-    fn wrap_err<M>(self, message: M) -> Result<T, Self::Context>
-    where
-        M: Message;
-
-    /// Lazily adds new contextual message to the [`Frame`] stack of a [`Report`].
-    ///
-    /// The function is only executed in the `Err` arm.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use core::fmt;
-    /// # use error::Result;
-    /// # fn load_resource(_: &User, _: &Resource) -> Result<()> { Ok(()) }
-    /// # struct User;
-    /// # struct Resource;
-    /// # impl fmt::Display for Resource { fn fmt(&self, _: &mut fmt::Formatter<'_>) -> fmt::Result { Ok(()) }}
-    /// use error::ResultExt;
-    ///
-    /// # let user = User;
-    /// # let resource = Resource;
-    /// # #[allow(unused_variables)]
-    /// let resource = load_resource(&user, &resource)
-    ///     .wrap_err_lazy(|| format!("Could not load resource {resource}"))?;
-    /// # Result::Ok(())
-    /// ```
-    fn wrap_err_lazy<M, F>(self, op: F) -> Result<T, Self::Context>
-    where
-        M: Message,
-        F: FnOnce() -> M;
-
-    /// Adds a context provider to the [`Frame`] stack of a [`Report`] returning
-    /// [`Result<T, Context>`]).
-    // TODO: come up with a decent example
-    fn provide_context<C>(self, context: C) -> Result<T, C>
-    where
-        C: Context;
-
-    /// Lazily adds a context provider to the [`Frame`] stack of a [`Report`] returning
-    /// [`Result<T, C>`]).
-    // TODO: come up with a decent example
-    fn provide_context_lazy<C, F>(self, op: F) -> Result<T, C>
-    where
-        C: Context,
-        F: FnOnce() -> C;
-}
 
 /// Iterator over the [`Frame`] stack of a [`Report`].
 ///
